move GDPR specs to SP specs
