+++ Example/ConsentViewController_ExampleTests/Extensions/SPURLExtensionsSpec.swift
